// src/types.ts
export enum Page {
  HOME = 'HOME',
  LAPORAN_BENCANA = 'LAPORAN_BENCANA',
  LAPORAN_KEGIATAN = 'LAPORAN_KEGIATAN',
  ABSEN_PIKET = 'ABSEN_PIKET',
  PENGUMUMAN = 'PENGUMUMAN',
  PANDUAN_DARURAT = 'PANDUAN_DARURAT',
  SUSUNAN_PENGURUS = 'SUSUNAN_PENGURUS',
  DATA_BASE_ANGGOTA = 'DATA_BASE_ANGGOTA',
}